requirements.txt
awsebcli==3.7.7
blessed==1.9.5
boto==2.42.0
botocore==1.4.41
cement==2.8.2
colorama==0.3.7
Django==1.9.7
django-allauth==0.25.2
django-storages-redux==1.3.2
django-summernote==0.8.5
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
jmespath==0.9.0
MySQL-python==1.2.5
oauthlib==1.1.2
pathspec==0.3.4
Pillow==3.3.0
python-dateutil==2.5.3
python-openid==2.2.5
pytz==2016.4
PyYAML==3.11
requests==2.9.1
requests-oauthlib==0.6.1
semantic-version==2.5.0
six==1.10.0
texttable==0.8.4
wcwidth==0.1.7
websocket-client==0.37.0
